requirements.txt
asgiref==3.5.2
Django==4.1.2
django-constance==2.9.1
django-node-assets==0.9.11
django-picklefield==3.1
django-qr-code==3.1.1
django-tables2==2.4.1
Pillow==9.2.0
psycopg2-binary==2.9
pydantic==1.10.2
pytz==2022.6
segno==1.5.2
sqlparse==0.4.3
typing_extensions==4.4.0
django-environ
django-storages
google-cloud-vision
google-cloud
google-cloud-bigquery
google-cloud-storage
gunicorn
django-widget-tweaks 
pandas
django-extensions
django-cors-headers
django-jazzmin